Updated Drawable interface. Now drawable objects store the reference to graphics inside them. Because of that, each drawable constructor will need graphics passed to it.
Updated the update in drawable method to receive the deltaTime parameter. Idk why but feel like this will be needed

Updated the seeds. Made it more OOP friendly (sorry Anton). Put the seeds list in the Seeds class. Now to check the seed intersection you need to get the list from the Seeds class. Intersection check is still in the GameWorld. (Think about a better way to check intersection)

Put the rectangle creation and the rectangle updates for the intersection check in the Drawable class. Now each drawable object has its own rectangle that updates position every tick. Use method getRectangle to get an updated version of the rectangle from the object. This helps with seed intersection check and saves memory because we do not create a new rectangle every time we check for a seed intersection.

Created a new animation sprite sheet for the pigeon.

Updated the animation class to utilize a multiple column sprites sheet for walking left and right.

Updated pigeon animation. Now it can walk left and right with variable animation speeds .

(For testing: Removed the timer for the loading screen for testing. Removed loosing when missing one seed for testing.)
All these updates seem to make the game work more stable. Game runs at 56-60fps. In order to fix the memory leak, fix the score counter. It created strings so the memory is used. Same goes for fps counter. Also the iterator for the seeds uses the memory to create the objects. Will check if can do something better.