Added ability to scale pixmap and rectangle
Pixmap can not be scaled in game. two methods were added to the AndroidPixmap class. One method accepts the scale factor. Meaning height and width will be scaled equally. The other method accepts height and width to which to scale.

Second change is the collision detection rectangle scaling.  Drawable class now has a field called rectangleScaleFactor. Whatever the scale factor is the rectangle will scale itself every time it is requested by a getRectangle method. Using this method the pegion collision detection rectangle is smaller. It is 85% of its original height.